requirements.txt
click==8.0.3
gitdb==4.0.9
GitPython==3.1.27
psutil==5.8.0
smmap==5.0.0
